pyproject.toml
[project]
name = "telegram-plans-bot"
version = "0.1.0"
description = "Telegram bot for managing plans and tasks"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "md2tgmd>=0.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
